+++ Game/main.py
@@ -11,9 +11,10 @@ This file contains important GUI code.
 Always run the game from this file.
 """
 from tkinter import *
-from multiprocessing import Process
-import connection
+import threading
 import battle
+import time
+import gc
 
 
 # Represents the graphical user interface of the game.
@@ -24,38 +25,58 @@ class GUI:
         """
         self.mainframe = None
         self.main = Tk()
-        self.pipe = None
-        self.online = False
         self.op = False
         self.game = False
-        self.ip = None
-        self.port = None
+        self.start = None
         self.process1 = None
         self.process2 = None
         self.process3 = None
         self.connect_process = None
+        self.name = StringVar(self.main)
+        self.name.set("Pick your username")
 
     def offgame(self):
         """
         Starts an offline game with a bot.
         """
-        self.process2 = Process(target=self.offgame_wrapper())
+        self.process2 = threading.Thread(target=self.offgame_wrapper())
         self.process2.start()
 
     def offgame_wrapper(self):
         """
         Wraps the function for starting an offline game with a bot.
         """
-        self.game = battle.LBattle()
-        # self.game.game_ui()
+        level = 1
+        self.main.withdraw()
+        self.game = battle.LBattle(level)
+        gamestatus = self.game.start()
+        while gamestatus == 0:
+            level += 1
+            gc.collect()
+            self.game = battle.LBattle(level)
+            gamestatus = self.game.start()
+
+        self.main.deiconify()
 
     def quit_game(self):
         """
         Quits the game and closes the GUI.
         """
         self.main.destroy()
+        self.main.quit()
 
-    def temp_name(self):
+    def checker(self):
+        while True:
+            try:
+                while self.name.get() == "Pick your username" or self.name.get() == "":
+                    self.start.config(state="disabled")
+                self.start.config(state="normal")
+                time.sleep(1)
+            except RuntimeError:
+                pass
+
+
+    def maingui(self):
         """
         Design the main title screen.
         """
@@ -65,8 +86,10 @@ class GUI:
         maintitle.pack(ipadx=20, ipady=10, expand=True)
         choice3 = Label(self.main, text="Play a Trainer!", font=("MS Comic Sans", "14"))
         choice3.pack(ipadx=20, ipady=20, expand=True)
-        start = Button(self.main, text="Play Bot!", command=self.offgame)
-        start.pack()
+        username = Entry(self.main, textvariable=self.name)
+        username.pack()
+        self.start = Button(self.main, text="Play Bot!", command=self.offgame)
+        self.start.pack()
         credit = Label(self.main, text="This program was made by Calvin, Ebaad and Josh", font=("MS Comic Sans", "10"))
         credit.pack(ipadx=20, ipady=20, expand=True)
         quit_button = Button(self.main, text="Quit", command=self.quit_game)
@@ -75,5 +98,7 @@ class GUI:
 
 
 if __name__ == "__main__":
-    b = GUI()
-    b.temp_name()
+    main = GUI()
+    side = threading.Thread(target=main.checker)
+    side.start()
+    main.maingui()