+++ Changelog.md
@@ -1,3 +1,16 @@
+# V.1.0.0
+## Alpha Release
+- Created main.py
+- Created enemies.py
+- Created player.py
+- Created JSON_Poke.py
+- Created Poke_Obj.py
+- Created battle.py
+- Added GUI
+- Added Pokemon objects
+- Added Player and Enemies objects
+- Added Battling system
+
 # V.0.0.0
 - Changelog Created
 - Finished Timeline